Separate signer proto.
