+++ core/proto/keytransparency_v1_types/keytransparency_v1_types.proto
@@ -22,6 +22,7 @@ syntax = "proto3";
 package keytransparency.v1.types;
 
 import "github.com/google/keytransparency/core/proto/ctmap/ctmap.proto";
+import "github.com/google/keytransparency/core/proto/signature/signature.proto";
 
 //
 // Data types.
@@ -91,7 +92,7 @@ message SignedKV {
   // signatures on key_value. Must be signed by keys from both previous and
   // current epochs. The first proves ownership of new epoch key, and the
   // second proves that the correct owner is making this change.
-  map<string, ctmap.DigitallySigned> signatures = 2;
+  map<string, signature.DigitallySigned> signatures = 2;
   // previous contains the hash of the previous entry that this mutation is
   // modifying creating a hash chain of all mutations. The hash used is
   // CommonJSON in "github.com/benlaurie/objecthash/go/objecthash".
