+++ core/proto/signature/gen.go
@@ -12,6 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//go:generate protoc -I=. -I=$GOPATH/src/ --go_out=:. ctmap.proto
+//go:generate protoc -I=. --go_out=:. signature.proto
 
-package ctmap
+package signature
