+++ core/proto/signature/signature.proto
@@ -1,0 +1,45 @@
+// Copyright 2016 Google Inc. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+syntax = "proto3";
+
+package signature;
+
+// DigitallySigned defines a way to digitally sign objects.
+message DigitallySigned {
+  // HashAlgorithm defines the approved methods for object hashing.
+  enum HashAlgorithm {
+    // No hash algorithm is used.
+    NONE = 0;
+    // SHA256 is used.
+    SHA256 = 4;
+    // SHA512 is used.
+    SHA512 = 6;
+  }
+
+  // SignatureAlgorithm defines the algorithm used to sign the object.
+  enum SignatureAlgorithm {
+    // Anonymous signature scheme.
+    ANONYMOUS = 0;
+    // ECDSA signature scheme.
+    ECDSA = 1;
+  }
+
+  // hash_algorithm contains the hash algorithm used.
+  HashAlgorithm hash_algorithm = 1;
+  // sig_algorithm contains the signing algorithm used.
+  SignatureAlgorithm sig_algorithm = 2;
+  // signature contains the object signature.
+  bytes signature = 3;
+}
+
